pyproject.toml
[tool.poetry]
name = "inky-dashboard"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
inky-dashboard = 'inky_dashboard.__main__:main'

[tool.poetry.dependencies]
python = "^3.11"
inky = {extras = ["rpi"], version = "^1.5.0"}
pillow = "^10.2.0"
playwright = "^1.41.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
